requirements.txt
PyQt6>=6.6.0
PyYAML>=6.0
markitdown[all]>=0.0.2
